Move notification.js under app/javascript/packs
