+++ app/views/layouts/application.html.erb
@@ -14,6 +14,7 @@
     <%-# stylesheet_pack_tag 'application', media: 'all'%>
     <%= javascript_include_tag 'application' %>
     <%= javascript_pack_tag 'application' %>
+    <%= javascript_pack_tag 'notification' %>
     <%= csrf_meta_tags %>
   </head>
   <body class="fixed-nav bg-light">
